refactor(formatter): Split `print/embed` file (#19699)
Simply extracted a portion of the file.
